restore schedule page feature flag
