+++ apps/site/src/app/schedule/page.tsx
@@ -1,25 +1,10 @@
-import { PortableText } from "@portabletext/react";
-import ClipboardSchedule from "./ClipboardSchedule";
-import { getSchedule } from "./getSchedule";
+import { Maintenance } from "@/views";
+import Schedule from "./schedule";
 
 export const revalidate = 60;
 
-import styles from "./page.module.scss";
+const Component = process.env.MAINTENANCE_MODE_SCHEDULE
+	? Maintenance
+	: Schedule;
 
-export default async function Schedule() {
-	const days = await getSchedule();
-
-	const schedule = days.map((events) =>
-		events.map(({ description, ...event }) => ({
-			...event,
-			description: <PortableText value={description} />,
-		})),
-	);
-
-	return (
-		<div className={styles.schedule}>
-			<h1>Schedule</h1>
-			<ClipboardSchedule schedule={schedule} />
-		</div>
-	);
-}
+export default Component;
